fix(macos): ContentPageView 키 명령을 리네임 상태에 맞게 조정
- KeyCommandView 표시 조건에서 isGridLayout 제거
- Enter/Return 키 입력을 renamingItemId로 가드해 리네임 중복 진입 방지
- 테스트 호환성을 위해 keyCode 리터럴에 UInt16 명시 변환

+++ apps/macos/Voyager/VoyagerTests/Pages/FileManager/Content/FileManagerContentKeyCommandHandlerTests.swift
@@ -218,7 +218,7 @@ final class FileManagerContentKeyCommandHandlerTests: XCTestCase {
 
     func testRenameShortcutStartsRenameForSingleSelectionInListAndGrid() async {
         for layout in [EntryViewLayoutState.Mode.list, .grid] {
-            for keyCode in [36, 76] {
+            for keyCode in [UInt16(36), UInt16(76)] {
                 let selected = makeEntry(name: "selected", fullPath: "/tmp/voyager/selected.txt")
 
                 var initialState = FileManagerContentState()
@@ -248,7 +248,7 @@ final class FileManagerContentKeyCommandHandlerTests: XCTestCase {
 
     func testRenameShortcutDoesNothingWhenRenameAlreadyActive() async {
         for layout in [EntryViewLayoutState.Mode.list, .grid] {
-            for keyCode in [36, 76] {
+            for keyCode in [UInt16(36), UInt16(76)] {
                 let selected = makeEntry(name: "selected", fullPath: "/tmp/voyager/selected.txt")
 
                 var initialState = FileManagerContentState()
@@ -272,7 +272,7 @@ final class FileManagerContentKeyCommandHandlerTests: XCTestCase {
 
     func testRenameShortcutDoesNothingForZeroOrMultipleSelection() async {
         for layout in [EntryViewLayoutState.Mode.list, .grid] {
-            for keyCode in [36, 76] {
+            for keyCode in [UInt16(36), UInt16(76)] {
                 var initialState = FileManagerContentState()
                 initialState.entryViewLayout.mode = layout
                 initialState.navigation.seedInitialFolderPath("/tmp/voyager")
